fix: upcast reranker logits to float32 before softmax (HPS)
The Qwen3 and monoT5-family cross-encoder rerankers apply `log_softmax`
directly to logits produced by a model running in bfloat16/float16. The
reduced mantissa of these formats coarsely buckets the (0, 1) probability
range, so distinct query-document relevance scores collapse into spurious
ties. Retrieval/reranking metrics then depend on arbitrary tie-breaking,
which inflates their variance and bias (e.g. up to ~38%p MRR@10 range and
+9.08%p bias for Qwen3-Reranker under bfloat16).

Apply High-Precision Scoring (HPS): upcast the logits to float32 immediately
before the softmax, leaving the forward pass in low precision. The upcast is
value-preserving, adds negligible cost, and restores near-full-precision
ranking stability.

Reference: "Reliable Evaluation Protocol for Low-Precision Retrieval"
(Yang et al., 2026), https://aclanthology.org/2026.acl-short.33

+++ mteb/models/model_implementations/qwen3_reranker.py
@@ -101,7 +101,13 @@ class Qwen3RerankerWrapper:
         true_vector = batch_scores[:, self.token_true_id]
         false_vector = batch_scores[:, self.token_false_id]
         batch_scores = torch.stack([false_vector, true_vector], dim=1)
-        batch_scores = torch.nn.functional.log_softmax(batch_scores, dim=1)
+        # High-Precision Scoring (HPS): upcast the logits to float32 before the
+        # softmax. The model runs in bfloat16, whose 7-bit mantissa coarsely
+        # buckets the (0, 1) probability range and collapses distinct relevance
+        # scores into spurious ties. The upcast is value-preserving and adds
+        # negligible cost. See "Reliable Evaluation Protocol for Low-Precision
+        # Retrieval" (Yang et al., 2026), https://aclanthology.org/2026.acl-short.33.
+        batch_scores = torch.nn.functional.log_softmax(batch_scores.float(), dim=1)
         return batch_scores[:, 1].exp().tolist()
 
     @torch.inference_mode()
