+++ mteb/models/model_implementations/rerankers_monot5_based.py
@@ -152,7 +152,14 @@ class MonoT5Reranker(RerankerWrapper):
         )
         batch_scores = output.scores[0]
         batch_scores = batch_scores[:, [self.token_false_id, self.token_true_id]]
-        batch_scores = torch.nn.functional.log_softmax(batch_scores, dim=1)
+        # High-Precision Scoring (HPS): upcast the logits to float32 before the
+        # softmax. Rerankers are commonly run in float16/bfloat16, whose reduced
+        # mantissa coarsely buckets the (0, 1) probability range and collapses
+        # distinct relevance scores into spurious ties. The upcast is
+        # value-preserving and adds negligible cost. See "Reliable Evaluation
+        # Protocol for Low-Precision Retrieval" (Yang et al., 2026),
+        # https://aclanthology.org/2026.acl-short.33.
+        batch_scores = torch.nn.functional.log_softmax(batch_scores.float(), dim=1)
         return batch_scores[:, 1].exp().tolist()
 
 
@@ -265,11 +272,14 @@ Relevant: """
             true_vector = batch_scores[:, self.token_true_id]
             false_vector = batch_scores[:, self.token_false_id]
             batch_scores = torch.stack([false_vector, true_vector], dim=1)
-            batch_scores = torch.nn.functional.log_softmax(batch_scores, dim=1)
+            # HPS: upcast to float32 before the softmax to avoid low-precision
+            # tie collisions (see the note in MonoT5Reranker.predict above).
+            batch_scores = torch.nn.functional.log_softmax(batch_scores.float(), dim=1)
             scores = batch_scores[:, 1].exp().tolist()
         else:
             batch_scores = self.model(**tokens).logits
-            batch_scores = torch.nn.functional.log_softmax(batch_scores, dim=1)
+            # HPS: upcast to float32 before the softmax (see note above).
+            batch_scores = torch.nn.functional.log_softmax(batch_scores.float(), dim=1)
             scores = batch_scores[:, 1].exp().tolist()
 
         return scores
